feat(mock): msw ts 마이그레이션
